refactor(expo-polyfills): build without `expo`
