+++ examples/medplum-react-native-example/package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@expo/webpack-config": "19.0.0",
     "@medplum/core": "2.1.10",
+    "@medplum/expo-polyfills": "*",
     "@medplum/react-hooks": "2.1.10",
     "expo": "49.0.13",
     "expo-status-bar": "1.7.1",
