+++ examples/medplum-react-native-example/src/App.tsx
@@ -19,7 +19,7 @@ const medplum = new MedplumClient({
 });
 
 // This is a module to get the Medplum client working on React Native by polyfilling a few Web APIs that are missing from the React Native runtime
-// On web, this is a no-op
+// On web, the polyfill function is loaded but nothing is replaced
 polyfillMedplumWebAPIs();
 
 export default function App(): JSX.Element {
