+++ packages/expo-polyfills/.gitignore
@@ -9,6 +9,7 @@ jsconfig.json
 # Xcode
 #
 build/
+dist/
 *.pbxuser
 !default.pbxuser
 *.mode1v3
