+++ packages/expo-polyfills/esbuild.mjs
@@ -1,0 +1,37 @@
+/* global console */
+/* eslint no-console: "off" */
+import esbuild from 'esbuild';
+import { nodeExternalsPlugin } from 'esbuild-node-externals';
+import { writeFileSync } from 'node:fs';
+
+const options = {
+  entryPoints: ['./src/index.ts'],
+  bundle: true,
+  platform: '',
+  loader: { '.ts': 'ts' },
+  resolveExtensions: ['.ts'],
+  target: 'es2021',
+  tsconfig: 'tsconfig.json',
+  minify: true,
+  sourcemap: true,
+  plugins: [nodeExternalsPlugin()],
+  external: ['expo-crypto', 'expo-standard-web-crypto'],
+};
+
+esbuild
+  .build({
+    ...options,
+    format: 'cjs',
+    outfile: './dist/cjs/index.cjs',
+  })
+  .then(() => writeFileSync('./dist/cjs/package.json', '{"type": "commonjs"}'))
+  .catch(console.error);
+
+esbuild
+  .build({
+    ...options,
+    format: 'esm',
+    outfile: './dist/esm/index.mjs',
+  })
+  .then(() => writeFileSync('./dist/esm/package.json', '{"type": "module"}'))
+  .catch(console.error);
