--- packages/expo-polyfills/expo-module.config.json
@@ -1,3 +1,0 @@
-{
-  "platforms": ["ios", "android", "web"]
-}
