--- packages/expo-polyfills/src/ExpoMedplumPolyfillsModule.web.ts
@@ -1,1 +1,0 @@
-export function polyfillMedplumWebAPIs(): void {}
